ENH: Split ICA into multiple steps
