+++ mne_bids_pipeline/steps/preprocessing/__init__.py
@@ -6,7 +6,8 @@ from . import (
     _03_maxfilter,
     _04_frequency_filter,
     _05_regress_artifact,
-    _06a_run_ica,
+    _06a1_fit_ica,
+    _06a2_find_ica_artifacts,
     _06b_run_ssp,
     _07_make_epochs,
     _08a_apply_ica,
@@ -20,7 +21,8 @@ _STEPS = (
     _03_maxfilter,
     _04_frequency_filter,
     _05_regress_artifact,
-    _06a_run_ica,
+    _06a1_fit_ica,
+    _06a2_find_ica_artifacts,
     _06b_run_ssp,
     _07_make_epochs,
     _08a_apply_ica,
